Revert "Feature/recurso - added search to recurso funcionallity"
