+++ core/services/RecursoService.ts
@@ -6,26 +6,14 @@ import { API } from "../../http/API";
 export default function RecursoService() {
 
     const [listaRecursos, setListaRecursos] = useState<Recurso[]>([]);
-    const [originalData, setOriginalData] = useState<Recurso[]>([]);
 
     useEffect(() => {
         API.get<Recurso[]>('Recurso').then((response) => {
             setListaRecursos(response.data);
-            setOriginalData(response.data);
-        }).catch((error) => {
-            console.log("Erro:::::::" + error);
-        });
-
-    }, []);
-
-    const getListaRecursos = async () => {
-        try {
-            const response = await API.get<Recurso[]>('Recurso');
-            setListaRecursos(response.data);
-        } catch (error) {
-            console.log(error);
         }
-    };
+
+        )
+    }, []);
 
     const deleteRecurso = async (id: number) => {
         try {
@@ -60,7 +48,7 @@ export default function RecursoService() {
         }
     };
 
-    return { listaRecursos, originalData, setOriginalData, setListaRecursos, deleteRecurso, saveRecurso, getListaRecursos };
+    return { listaRecursos, deleteRecurso, saveRecurso };
 
 }
