+++ models/Recurso.ts
@@ -8,4 +8,5 @@ export type Recurso = {
     dataCadastro: string;
     status: number;
     usuarioId: string;
+    usuario: Usuario;
 }
