+++ package.json
@@ -10,7 +10,7 @@
     "@shopify/flash-list": "1.4.0",
     "@types/react": "~18.0.27",
     "axios": "^1.3.5",
-    "expo": "^48.0.17",
+    "expo": "^48.0.7",
     "expo-blur": "~12.2.2",
     "expo-constants": "~14.2.1",
     "expo-document-picker": "^11.2.2",
@@ -33,9 +33,7 @@
     "react-native-screens": "~3.20.0",
     "react-native-svg": "13.4.0",
     "react-native-web": "~0.18.7",
-    "typescript": "^4.9.4",
-    "uuid": "^9.0.0",
-    "uuid-random": "^1.3.2"
+    "typescript": "^4.9.4"
   },
   "devDependencies": {
     "@babel/core": "^7.19.3",
